+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Node.getParent() returns None if the node has no parent
 - add NULL pointer checks to all Python wrapper classes
 - add Event.getRow() and Row.name
+- expose domain changes, bound changes, branching decisions, added constraints for nodes
 - define Python object identity based on underlying SCIP object pointers, so that e.g. rows and columns can be added to sets, and testing for equality is consistent over time.
 
 ## 2.2.3 - 2019-12-10